+Fonction à ajouter sur l'interface front/back -Modification struct trace avec ajout de la visibilité +Ajout visibilité du trajet -Modification struct pt_interet avec importance, plage horaire et la position

+++ Code/headers/trace.h
@@ -3,7 +3,7 @@
 #include "structures.h"
 
 trace * createTrace(void);
-void initTrace(trace * parT, double parDate, point parCoord, trace * parPrec, trace * parSuiv);
+void initTrace(trace * parT, long parDate, point parCoord, trace * parPrec, trace * parSuiv);
 trajet * initTrajet(void);
 void ajoutTrace(trajet * parTr, trace * parT);
 void afficheTrace(trace * parT);
